fix implementation oversight and 1 bug
implementation oversight that prevents the player from dying when
getting snatched into a deadly room or prevents getting snatched
multiple times. Player has a 2/20 chance of getting snatched again and a
3/20 chance of dying everytime the player is snatched.

one subtle bug fixedd where if the player plays again with the same
setup the random number generator produces the same results because I
had it saved as a static readonly field.